Hotfix: Simulator endpoints now works.
The correct ports have been exposed in the docker-compose file.
The correct path is now set in all simulator controllers.
The test_sim_api script, have been modified so its possible
to test both in development and in production.
The README file have been updated with usage instrucitons
on how to use the simulator test script.