Chore/#53 create a deploy script (#64)
* ran npm install on missing packages

* Added appsettings to .dockerignore

* Initial deploy.sh script created.

The script is not yet working, since it is missing
the names of the docker images to pull and run.

* Added container_name attribute in composefile.

Added a container_name attribute and value, in the docker compose file,
for the minitwit-api.

* Not in a working state.

* Changed docker run in 'deploy.sh'.

Correct exposure of ports when running api container with docker.
Added enviroment variables to when running web container with docker.

* Deploy script is now working.

* Added enviroment variables to scripts.

Added enviroment variables to the docker compose file,
and the deploy script.